00_Projects/01_Deep_Research_Agent/pyproject.toml
[project]
name = "01-deep-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.8",
    "tavily-python>=0.7.10",
]
